feature(itchat): add UOS weixin desktop patch
It turns out that `wechat-uos` provided by Tencent still uses the web
interface, and someone finds the way to utilize the `wechat-uos` to
login into the web for the ones who cannot login by web. This commit
uses the information provided by the others. It does the three basic
things:

1. Change itchat `config.py` to add the corresponding version and
   `UOS_PATCH_EXTSPAM`.
2. Change itchat `components/login.py` to add the latest weixin check,
   `skey` field and `pass_ticket` field.
3. In order to get the message from weixin, eveytime the device ID
   should be changed, so this commit uses random string to generate
   the device ID.

However, this patch doesn't work for WeChat users.